Add hops/fermentable, store neural network, fix gravity normalisation bug
Added hops and fermentables to Beerbrain. It uses the relative ratio of
ibu contribution of each hops in relation to the total ibu (tinseth).

Added fermentables to Beerbrain. Relative amounts of fermentables in
ratio to total amount is used for each fermentable.

Each type of ingredient thus is represented as a feature in the neural
network input and is given a value between zero and 1.

Fixed bug were gravity normalisation returned negative values.