Refactor, separate training and classification testing
Major refactor of neural network code. Moved all the training and
classification logic into a separate brain.coffee module.

Separated training and classification logic into two separate
functions. brain.train() trains the neural network based off 80% of the
training set, while brain.classify() runs a classification test on 20%
of the dataset.

The brain.train() function shuffles the cases (recipes), draws the 80%
first, and stores the rest in a file. This file is read by the
classify() function. This ensures a random selection of cases to use
for training and classification, and no overlap between the two.